# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Session cookie (browser) or Bearer token (scripts, mobile). Cookie-authenticated POST/PUT/PATCH/DELETE must send X-Rigel-Client.",
        "title": "RigelLedger API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/admin/access-requests": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Requests for an account (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pending, approved or rejected",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.AccessRequestDTO"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/access-requests/{requestID}/approve": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Approve a request: the person is invited (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "request id",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/admin/access-requests/{requestID}/reject": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Reject a request (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "request id",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/admin/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Recent sign-in and security events of everyone (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "at most 500",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.AuthEventDTO"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/invitations": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Invite a user by username and email (admin)",
                "parameters": [
                    {
                        "description": "who",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.inviteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminUserDTO"
                        }
                    },
                    "409": {
                        "description": "taken, or mail_failed (the invitation is saved; resend it)",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/admin/invitations/{userID}": {
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "Withdraw an invitation that was not accepted; deletes that user (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "invited user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/admin/invitations/{userID}/resend": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Mail a fresh invitation link; the old one stops working (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "invited user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/admin/mail": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Outgoing mail settings (admin); the password is stored sealed and never returned",
                "parameters": [
                    {
                        "description": "mail",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.adminMailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminSettingsDTO"
                        }
                    }
                }
            }
        },
        "/api/admin/mail/test": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Send a test email to yourself (admin)",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "409": {
                        "description": "mail_failed or mail_disabled",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/admin/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "System settings (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminSettingsDTO"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Replace registration, security and default settings (admin)",
                "parameters": [
                    {
                        "description": "every field",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.adminSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminSettingsDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/admin/users/{userID}": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Activate, deactivate, promote or demote another user (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "changes",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.adminUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminUserDTO"
                        }
                    },
                    "409": {
                        "description": "last_admin",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        },
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "routes.AccessRequestDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "decided_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "routes.AdminSettingsDTO": {
            "type": "object",
            "properties": {
                "default_date_format": {
                    "type": "string"
                },
                "default_display_currency": {
                    "type": "string"
                },
                "default_language": {
                    "type": "string"
                },
                "default_theme": {
                    "type": "string"
                },
                "default_timezone": {
                    "type": "string"
                },
                "invite_ttl_seconds": {
                    "type": "integer"
                },
                "login_ip_max_failures": {
                    "type": "integer"
                },
                "login_max_failures": {
                    "type": "integer"
                },
                "login_user_max_failures": {
                    "type": "integer"
                },
                "login_window_seconds": {
                    "type": "integer"
                },
                "mail_configured": {
                    "type": "boolean"
                },
                "mail_driver": {
                    "type": "string"
                },
                "mail_from": {
                    "type": "string"
                },
                "mail_from_name": {
                    "type": "string"
                },
                "mfa_methods": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mfa_required": {
                    "type": "boolean"
                },
                "registration": {
                    "type": "string"
                },
                "session_ttl_seconds": {
                    "description": "null: SESSION_TTL from the environment.",
                    "type": "integer"
                },
                "smtp_host": {
                    "type": "string"
                },
                "smtp_password_set": {
                    "type": "boolean"
                },
                "smtp_port": {
                    "type": "integer"
                },
                "smtp_security": {
                    "type": "string"
                },
                "smtp_user": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "routes.AdminUserDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "initialized": {
                    "type": "boolean"
                },
                "invite_pending": {
                    "type": "boolean"
                },
                "invited": {
                    "description": "has not accepted yet",
                    "type": "boolean"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "routes.AuthEventDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "detail": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "event": {
                    "type": "string"
                },
                "failure": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "routes.adminMailRequest": {
            "type": "object",
            "properties": {
                "clear_smtp_password": {
                    "type": "boolean"
                },
                "mail_driver": {
                    "type": "string"
                },
                "mail_from": {
                    "type": "string"
                },
                "mail_from_name": {
                    "type": "string"
                },
                "smtp_host": {
                    "type": "string"
                },
                "smtp_password": {
                    "description": "Empty or absent keeps the stored password.",
                    "type": "string"
                },
                "smtp_port": {
                    "type": "integer"
                },
                "smtp_security": {
                    "type": "string"
                },
                "smtp_user": {
                    "type": "string"
                }
            }
        },
...
        "routes.adminUserRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "is_admin": {
                    "type": "boolean"
                }
            }
        },
...
        "routes.inviteRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}